requirements.txt
keras==2.12.0
lmdb==1.4.1
numpy==1.23.5
matplotlib==3.7.1
Pillow==9.5.0
tensorflow==2.12.0
scipy==1.10.1